Support setting URI SANs in CSRs
In #70 support was added to the Vault issuer to allow setting URI SANs.
The particular fix leaves two things to be desired:

(1) It is Vault-specific
(2) It interacts with other configuration options in an unfortunate way.

While (1) is self-explanatory, (2) deserves more explanation.

With Vault, you have a binary knob: either take SANs from the
CSR or from the Vault API request. In order to set URI SANs
with the support in #70, one can configure Vault to take all SANs
from the API request. But if you do this you lose the ability to set
DNS SANs, since this functionality is only supported in Certify through
the base certify.CertConfig which ends up setting SANs on the CSR itself
(which is now being ignored by Vault). If you do not set this flag in
Vault the URI SANs specified through the vault.Issuer are ignored.

To summarize, there is no way with any combination of Certify
configuration and Vault configuration to specify both URI SANs
and DNS SANs.

This commit addresses that limitation, by supporting setting URI SANs
directly on the CSR. This has been supported since Go 1.10.